+++ src/game_chars_alien.c
@@ -6,7 +6,7 @@
 /*   By: nmikuka <nmikuka@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/03 23:28:13 by psmolin           #+#    #+#             */
-/*   Updated: 2025/11/12 15:02:55 by nmikuka          ###   ########.fr       */
+/*   Updated: 2025/11/12 20:02:25 by nmikuka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,13 +15,13 @@
 static bool	ft_can_see_player(t_vec2 char_pos, t_vec2 player_pos)
 {
 	t_rayrender	ray;
-	float		wall_distance;
-	float		player_distance;
+	double		wall_distance;
+	double		player_distance;
 
-	char_pos.x += 0.5f;
-	char_pos.y += 0.5f;
-	player_pos.x += 0.5f;
-	player_pos.y += 0.5f;
+	char_pos.x += 0.5;
+	char_pos.y += 0.5;
+	player_pos.x += 0.5;
+	player_pos.y += 0.5;
 	ray.start = char_pos;
 	ray.dir = ft_normalize_vec2((t_vec2){
 			player_pos.x - char_pos.x,
@@ -44,7 +44,7 @@ void	ft_ai_alien_moving(t_char *ch)
 	{
 		ch->ai.timer -= ft_game()->dt;
 		ch->ai.lerp_progress += ch->ai.lerp_speed * ft_game()->dt;
-		if (ch->ai.lerp_progress >= 1.0f || ch->ai.timer <= 0.0f)
+		if (ch->ai.lerp_progress >= 1.0 || ch->ai.timer <= 0.0)
 			ch->ai.state = AI_IDLE;
 		ch->sprite.pos = ft_lerpvec2(ch->ai.start_pos,
 				ch->ai.target_pos, ch->ai.lerp_progress);
@@ -53,19 +53,19 @@ void	ft_ai_alien_moving(t_char *ch)
 
 void	ft_ai_alien(t_char *ch)
 {
-	float	distance;
+	double	distance;
 
 	if (ch->ai.state == AI_IDLE)
 	{
 		if (ft_can_see_player(ch->sprite.pos, ft_game()->player->pos))
 		{
 			ch->ai.state = AI_MOVING;
-			ch->ai.lerp_progress = 0.0f;
+			ch->ai.lerp_progress = 0.0;
 			ch->ai.speed = AI_SPEED_ALIEN;
 			distance = ft_vec2_length((t_vec2){
 					ft_game()->player->pos.x - ch->sprite.pos.x,
 					ft_game()->player->pos.y - ch->sprite.pos.y});
-			if (distance > 0.01f)
+			if (distance > 0.01)
 				ch->ai.lerp_speed = ch->ai.speed / distance;
 			ch->ai.start_pos = ch->sprite.pos;
 			ch->ai.target_pos = ft_game()->player->pos;
