+++ src/init_data.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   init_data.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: psmolin <psmolin@student.42heilbronn.de    +#+  +:+       +#+        */
+/*   By: nmikuka <nmikuka@student.42heilbronn.de    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/26 13:45:35 by psmolin           #+#    #+#             */
-/*   Updated: 2025/11/11 13:27:02 by psmolin          ###   ########.fr       */
+/*   Updated: 2025/11/12 20:02:02 by nmikuka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ static void	ft_init_render(void)
 
 	game = ft_game();
 	game->render.projection_plane_dist = (WIDTH / 2.0) / tan(FOV_RAD / 2.0);
-	game->render.depth = malloc(sizeof(float) * (WIDTH / PIXEL_SIZE + 1));
+	game->render.depth = malloc(sizeof(double) * (WIDTH / PIXEL_SIZE + 1));
 	if (!game->render.depth)
 		ft_exit_perror("Could not allocate memory for render depth buffer\n");
 }
@@ -48,16 +48,16 @@ static void	ft_init_player(void)
 	player = malloc(sizeof(t_player));
 	if (!player)
 		ft_exit_perror("Could not allocate memory for player\n");
-	player->hp = 100.0f;
+	player->hp = 100.0;
 	player->pos.x = 0;
 	player->pos.y = 0;
-	player->lookdir.x = 1.0f;
-	player->lookdir.y = 0.0f;
+	player->lookdir.x = 1.0;
+	player->lookdir.y = 0.0;
 	player->mov_control = (t_point){0, 0};
 	player->rot_control = (t_point){0, 0};
-	player->mouse_pos = (t_vec2){-1.0f, -1.0f};
-	player->mouse_diff = (t_vec2){0.0f, 0.0f};
-	player->lookupdown = 0.0f;
+	player->mouse_pos = (t_vec2){-1.0, -1.0};
+	player->mouse_diff = (t_vec2){0.0, 0.0};
+	player->lookupdown = 0.0;
 	player->is_jumping = false;
 	player->is_shaking = false;
 	ft_game()->player = player;
@@ -87,8 +87,8 @@ static void	ft_init_game(void)
 	game->mmap.enabled = false;
 	game->mmap.opening = false;
 	game->mmap.picked = false;
-	game->mmap.lerp_progress = 0.0f;
-	game->mmap.lerp_speed = 10.0f;
+	game->mmap.lerp_progress = 0.0;
+	game->mmap.lerp_speed = 10.0;
 }
 
 void	ft_initialize(void)
